TYPE: feature
SUBJECT: Firebase user authentication applied + HTML form replaced with react-hook-form

DESCRIPTION: Firebase authentication with email/password applied. Still not able to store the name of the user, though. The retrieved user object is yet to be stored in Context and LocalStorage. The HTML form element in userForm now uses react-hook-form.