TYPE: feature-improvement
SUBJECT: Signout functionality implemented. Routing based on user logged-in status.

DESCRIPTION: Signout functionality completely implemented. Didn't use local storage as Firebase keeps the track of logged-in user with auth object. Routing implemented: default page set to login if the user is not logged-in, default page is home if the user is logged-in.